Changes to naming methods for tests, uncommented some tests that we need in ScheduleServiceTests

+++ SportCenter-Backend/src/main/java/ca/mcgill/ecse321/Sport/Center/Application/ECSE321/service/SchedulingService.java
@@ -148,7 +148,10 @@ public class SchedulingService {
      * @author Pei Yan
      */
     @Transactional
-    public void deleteSession(int sessionId){
+    public void deleteSession(int sessionId) throws Exception {
+        if (!sessionRepository.existsById(sessionId)) {
+            throw new Exception("No session with given ID");
+        }
         sessionRegistrationRepository.deleteAllBySessionId(sessionId);
         
         // Delete session from repository
